Create new components for `<ManageItemList/>` and `<PopupItemList/>`
Their Redux containers have gotten sufficiently-complex that it was time to
create new React components to encapsulate this stuff.
