+++ src/webextension/list/manage/containers/all-items.js
@@ -2,38 +2,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Localized } from "fluent-react";
-import PropTypes from "prop-types";
-import React from "react";
 import { connect } from "react-redux";
 
+import ManageItemList from "../components/manage-item-list";
 import { requestSelectItem } from "../../actions";
 import { parseFilterString, filterItem } from "../../filter";
 import { NEW_ITEM_ID } from "../../common";
-import ItemList, { ItemListPlaceholder } from "../../components/item-list";
 
 const collator = new Intl.Collator();
-
-function AllItems({totalItemCount, ...props}) {
-  if (props.items.length === 0) {
-    return (
-      <Localized id={`all-items-${totalItemCount ? "no-results" :
-                                                   "get-started"}`}>
-        <ItemListPlaceholder>
-          wHEn yOu cREATe an eNTRy...
-        </ItemListPlaceholder>
-      </Localized>
-    );
-  }
-  return (
-    <ItemList {...props}/>
-  );
-}
-
-AllItems.propTypes = {
-  totalItemCount: PropTypes.number.isRequired,
-  ...ItemList.propTypes,
-};
 
 export default connect(
   (state, ownProps) => {
@@ -52,4 +28,4 @@ export default connect(
   (dispatch) => ({
     onChange: (id) => dispatch(requestSelectItem(id)),
   }),
-)(AllItems);
+)(ManageItemList);
