+++ src/webextension/list/popup/components/popup-item-list.css
@@ -2,12 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-.all-items-container {
+.item-list-container {
   display: flex;
   flex-direction: column;
   overflow: hidden;
 }
 
-.all-items {
+.item-list {
   flex: 1;
 }
